+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Debian 10 (buster) support
 
+### Changes
+
+- Updated cronjob
+- Improve module cloning duration by adding `--depth=50` to `git clone`
+
 # [3.6.3] - 2019-07-15
 
 ### Added